Update metadata service
Update the way the metadata service was trying to identify the hostname
of the Virtual Machine. Instead of doing this with reverse DNS lookup,
we now try to find the information in the foreman.hosts hash.
The DNS way was buggy due to the ubuntu cloud image performing the first
DHCP request with 'ubuntu' hostname.